Consolidate lookup table logic
